pyproject.toml
[project]
name = "mini-market-analyzer"
version = "0.1.0"
description = "A CLI tool for market analysis using technical indicators and Gemini 2.5 Flash."
readme = "docs/overview.md"
requires-python = ">=3.12"
dependencies = [
    "pandas",
    "pandas-ta",
    "yfinance",
    "google-genai",
    "typer",
    "plotext",
    "rich",
    "python-dotenv>=1.2.1",
    "prompt-toolkit>=3.0.52",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mini_market_analyzer"]

[dependency-groups]
dev = [
    "pytest",
    "ruff",
    "mypy",
    "types-requests",
    "types-setuptools",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "SIM", "UP", "PL", "RUF"]
ignore = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["PLR2004"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "requires_api_key: marks tests that require API keys (auto-skipped if not set)",
]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
